--- src/canvas/ElementsManager/types.ts
@@ -1,9 +1,0 @@
-export type TCanvasSizeOptions = {
-  backstoreOnly?: boolean;
-  cssOnly?: boolean;
-};
-
-export type CanvasItem = {
-  el: HTMLCanvasElement;
-  ctx: CanvasRenderingContext2D;
-};
